refactor: split app-log backend into process/stream/platform modules (#161)
